requirements.txt
playwright
pandas
tqdm
openpyxl
requests
python-dotenv